pyproject.toml
[project]
name = "magi-server"
version = "0.1.0"
description = "MAGI model inference server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "transformers>=4.35.0",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "pillow>=10.0.0",
  "numpy>=1.24.0",
  "pydantic>=2.5.0",
  "einops>=0.8.1",
  "shapely>=2.1.1",
  "matplotlib>=3.10.3",
  "timm>=1.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
